requirements.txt
google-genai>=0.10.0
pandas>=2.0.0
numpy>=1.25.0
python-dotenv>=1.0.0
plotly>=6.0.0
dash>=3.0.0